Move CollectionDialog to AC
