+++ apps/web/src/app/admin-console/organizations/collections/vault.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from "@angular/core";
 
 import { LooseComponentsModule } from "../../../shared/loose-components.module";
 import { SharedModule } from "../../../shared/shared.module";
-import { CollectionDialogComponent } from "../../../vault/components/collection-dialog";
 import { OrganizationBadgeModule } from "../../../vault/individual-vault/organization-badge/organization-badge.module";
 import { ViewComponent } from "../../../vault/individual-vault/view.component";
+import { CollectionDialogComponent } from "../shared/components/collection-dialog";
 
 import { CollectionBadgeModule } from "./collection-badge/collection-badge.module";
 import { GroupBadgeModule } from "./group-badge/group-badge.module";
