+++ apps/web/src/app/admin-console/organizations/shared/components/collection-dialog/index.ts
