+++ apps/web/src/app/vault/individual-vault/vault.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from "@angular/core";
 
 import { CollectionBadgeModule } from "../../admin-console/organizations/collections/collection-badge/collection-badge.module";
 import { GroupBadgeModule } from "../../admin-console/organizations/collections/group-badge/group-badge.module";
+import { CollectionDialogComponent } from "../../admin-console/organizations/shared/components/collection-dialog";
 import { LooseComponentsModule, SharedModule } from "../../shared";
-import { CollectionDialogComponent } from "../components/collection-dialog";
 
 import { BulkDialogsModule } from "./bulk-action-dialogs/bulk-dialogs.module";
 import { OrganizationBadgeModule } from "./organization-badge/organization-badge.module";
